requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
validators==0.20.0
gunicorn==21.2.0
psycopg2-binary==2.9.9 